working on my model

+++ backend/main.py
@@ -1,5 +1,4 @@
 import os
-
 from fastapi import FastAPI, UploadFile, File
 from starlette.middleware.cors import CORSMiddleware
 from starlette.responses import JSONResponse
@@ -9,12 +8,13 @@ app = FastAPI()
 
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=["http://localhost:5173"],
+    allow_origins=["http://localhost:5173"],  # Dostosuj do swoich potrzeb
     allow_credentials=True,
     allow_methods=["*"],
     allow_headers=["*"],
 )
 
+# Ładowanie modelu podczas startu aplikacji
 model = load_model()
 
 @app.get("/")
@@ -24,12 +24,19 @@ async def root():
 @app.post("/predict")
 async def predict(image: UploadFile = File(...)):
     try:
+        # Sprawdzenie czy przesłany plik jest obrazem
+        if not image.content_type.startswith("image/"):
+            return JSONResponse({"error": "Plik musi być obrazem."}, status_code=400)
+
+        # Ścieżka do tymczasowego pliku
         temp_file_path = f"./temp_{image.filename}"
         with open(temp_file_path, "wb") as temp_file:
-            temp_file.write(image.file.read())
+            temp_file.write(await image.read())
 
+        # Predykcja
         prediction = predict_image(temp_file_path, model)
 
+        # Usunięcie tymczasowego pliku
         os.remove(temp_file_path)
 
         return JSONResponse(prediction)
